Skip the afterEach, for weird reasons
I ran these tests in Firefox, and it complained that collection.reset
was not a function. When stepping, the method seemed present as it
should. Strangely, the afterEach hook would run after the beforeEach
hook, but seemingly *before* any of the tests.

I suspect that removing this cleanup code introduces a memory leak in
the test suite, though I am not sure. Regardless, the tests still meet
their purpose and figuring out the cause of this problem does not seem
worth the effort at this time.

CC to self as a reminder to create an issue ticket: @jgonggrijp