Outfactor alt-click debugging tool from readit-interface
I had to give up on providing a ready-to-use constructor. You *cannot*
use `super()` in a classic scenario because it forces users to
transpile, but you *must* use `super()` in a modern scenario because
you are not allowed to call the constructor without `new`. However,
you can *only* use `super()` inside a class defined using the `class`
keyword, so that would force users to transpile!